*: move make targets around such that release-related targets are in their own Makefile, ex. image builds are in release/Makefile
images/: move all Dockerfiles from hack/image to here, update references

hack/: remove unnecessary wrapper scripts
