+++ images/custom-scorecard-tests/Dockerfile
@@ -1,16 +1,17 @@
 # Base image
 FROM registry.access.redhat.com/ubi8/ubi-minimal:latest
 
-ENV TEST=/usr/local/bin/custom-scorecard-tests \
-    USER_UID=1001 \
-    USER_NAME=test
+ENV HOME=/opt/custom-scorecard-tests \
+    USER_NAME=custom-scorecard-tests \
+    USER_UID=1001
 
-# install test binary 
-COPY custom-scorecard-tests ${TEST}
+RUN echo "${USER_NAME}:x:${USER_UID}:0:${USER_NAME} user:${HOME}:/sbin/nologin" >> /etc/passwd
 
-COPY bin /usr/local/bin
-RUN  /usr/local/bin/user_setup
+WORKDIR ${HOME}
 
-ENTRYPOINT ["/usr/local/bin/entrypoint"]
+ARG BIN=custom-scorecard-tests
+COPY $BIN /usr/local/bin/custom-scorecard-tests
+
+ENTRYPOINT ["/usr/local/bin/custom-scorecard-tests"]
 
 USER ${USER_UID}
