+++ images/custom-scorecard-tests/main.go
