+++ images/scorecard-test-kuttl/Dockerfile
@@ -1,18 +1,18 @@
-# Base image
 #FROM docker.io/kudobuilder/kuttl@sha256:c9c5edc27ba6e5e994ffd8cea03368c0d4e274f3c7a00f51c1e360feb573f24e
 FROM kudobuilder/kuttl:v0.6.1
 
-ENV TESTKUTTL=/usr/local/bin/scorecard-test-kuttl \
+ENV HOME=/opt/scorecard-test-kuttl \
+    USER_NAME=scorecard-test-kuttl \
     USER_UID=1001 \
-    USER_NAME=test \
     KUBECONFIG=
 
-# install test binary 
-COPY scorecard-test-kuttl ${TESTKUTTL}
+RUN echo "${USER_NAME}:x:${USER_UID}:0:${USER_NAME} user:${HOME}:/sbin/nologin" >> /etc/passwd
 
-COPY bin /usr/local/bin
-RUN  /usr/local/bin/user_setup
+WORKDIR ${HOME}
 
-ENTRYPOINT ["/usr/local/bin/entrypoint"]
+ARG BIN=scorecard-test-kuttl
+COPY $BIN /usr/local/bin/scorecard-test-kuttl
+
+ENTRYPOINT ["/usr/local/bin/scorecard-test-kuttl"]
 
 USER ${USER_UID}
