+++ images/scorecard-test-kuttl/main.go
