+++ release/changelog/internal/testdata/invalid_yaml/fragment1.yaml
