+++ test/e2e-helm/e2e_helm_scorecard_test.go
@@ -41,7 +41,7 @@ var _ = Describe("Testing Helm Projects with Scorecard", func() {
 			runScorecardCmd := exec.Command(tc.BinaryName, "scorecard", "bundle",
 				"--selector=suite=basic",
 				"--output=json",
-				"--wait-time=40s")
+				"--wait-time=60s")
 			scorecardOutputBytes, err := tc.Run(runScorecardCmd)
 			Expect(err).NotTo(HaveOccurred())
 			err = json.Unmarshal(scorecardOutputBytes, &scorecardOutput)
@@ -53,7 +53,7 @@ var _ = Describe("Testing Helm Projects with Scorecard", func() {
 			runOLMScorecardCmd := exec.Command(tc.BinaryName, "scorecard", "bundle",
 				"--selector=suite=olm",
 				"--output=json",
-				"--wait-time=40s")
+				"--wait-time=60s")
 			scorecardOutputBytes, err = tc.Run(runOLMScorecardCmd)
 			Expect(err).To(HaveOccurred())
 			err = json.Unmarshal(scorecardOutputBytes, &scorecardOutput)
